+++ SquidSquAds/src/main/java/com/squidsquads/form/campaign/request/CreateRequest.java
@@ -6,7 +6,7 @@ public class CreateRequest {
 
     private String name;
     private String redirectUrl;
-    private Long[] profileIds;
+    private Integer[] profileIds;
     private String startDate;
     private String endDate;
     private BigDecimal budget;
@@ -22,7 +22,7 @@ public class CreateRequest {
         return redirectUrl;
     }
 
-    public Long[] getProfileIds() {
+    public Integer[] getProfileIds() {
         return profileIds;
     }
 
@@ -59,7 +59,7 @@ public class CreateRequest {
         this.redirectUrl = redirectUrl;
     }
 
-    public void setProfileIds(Long[] profileIds) {
+    public void setProfileIds(Integer[] profileIds) {
         this.profileIds = profileIds;
     }
