+++ SquidSquAds/src/main/java/com/squidsquads/form/campaign/response/ListResponseItem.java
@@ -2,17 +2,17 @@ package com.squidsquads.form.campaign.response;
 
 public class ListResponseItem {
 
-    private final Long id;
+    private final Integer id;
     private final String name;
     private final String dateCreated;
 
-    public ListResponseItem(Long id, String name, String dateCreated) {
+    public ListResponseItem(Integer id, String name, String dateCreated) {
         this.id = id;
         this.name = name;
         this.dateCreated = dateCreated;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
